create workout added to home navigation

+++ app/src/main/java/com/example/hangboard/workout/dto/Activity.kt
@@ -1,4 +1,4 @@
 package com.example.hangboard.workout.dto
 
 
-data class Activity(val name: String, val rest: Long, val exercises: List<Exercise>)
+data class Activity(val name: String, val rest: Int, val exercises: List<Exercise>)
